GroceryList/requirements.txt
Flask==2.2.5
flask_sqlalchemy==3.1.1
openai==1.30.1
Pillow==10.1.0
Pillow==10.3.0
python-dotenv==1.0.1
Requests==2.31.0
